removing some older references, breaking stuff into diff class files
