+++ src/modules/cmdpal/src/Plugins/AllApps/Programs/IShellLinkHelper.cs
@@ -2,7 +2,7 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Wox.Infrastructure;
+namespace AllApps;
 
 public interface IShellLinkHelper
 {
