+++ src/modules/cmdpal/src/Plugins/AllApps/Programs/Win32Program.cs
@@ -5,20 +5,12 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
 using System.Security;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-using System.Windows.Input;
 // using Microsoft.Plugin.Program.Logger;
 // using Microsoft.Plugin.Program.Utils;
 using Microsoft.Win32;
- using Wox.Infrastructure;
-using Wox.Infrastructure.FileSystemHelper;
 // using Wox.Plugin;
 // using Wox.Plugin.Logger;
 // using DirectoryWrapper = Wox.Infrastructure.FileSystemHelper.DirectoryWrapper;
